requirements.txt
numpy
sympy
requests
colorama
python-twitter
